requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL driver

# Authentication
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
